Mavenize the project, drop Permissions support (untested)
