--- src/me/timvisee/WorldPortal/WorldPortalEntityListener.java
@@ -1,13 +1,0 @@
-package me.timvisee.WorldPortal;
-
-import org.bukkit.event.Listener;
-
-public class WorldPortalEntityListener implements Listener {
-	public static WorldPortal plugin;
-
-	public WorldPortalEntityListener(WorldPortal instance) {
-		plugin = instance;
-	}
-	
-	// Create function to teleport mobs when they use a portal
-}
